feature: Improved Data Type conversion.
This refactors the data type logic in tedious, making it more strict and
more accurate.

If the parameter specified by the user can not be converted to the data
type of the parameter without a loss of precision, the request will not
be executed and an error will be passed to the request's callback
instead.

The same parameter value conversion logic is now also used for both
regular requests as well as for bulk loads, fixing a long standing issue
where storing the same data via these two methods would result either
result in a failure or different data being stored in the database.